pyproject.toml
[tool.poetry]
name = "govtext-client"
version = "0.1.0"
description = "CLI and SDK for GovText"
authors = [
  "Charlton Lim <charlton@dsaid.gov.sg>"
]

[tool.poetry.dependencies]
click = "^7.1.2"
marshmallow = "^3.6.0"
python = "^3.7"
pyyaml = "^5.3.1"
requests = "^2.23.0"
tableprint = "^0.9.0"
wasabi = "^0.6.0"

[tool.poetry.dev-dependencies]
bandit = "^1.6.2"
black = "^19.10b0"
factory_boy = "^2.12.0"
flake8 = "^3.7.9"
isort = "^4.3.21"
mypy = "^0.761"
pre-commit = "^2.1.0"
pytest = "^5.3.5"
pytest-dotenv = "^0.4.0"
pytest-mock = "^2.0.0"
sqlalchemy-stubs = "^0.3"

[tool.poetry.scripts]
govtext = "govtext_client.cli.govtext:govtext"

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
atomic = true
skip_glob = ["*/setup.py"]
known_first_party = [
  "govtext_client",
]
known_third_party = [
  "factory",
  "marshmallow",
  "pytest",
  "pytest_mock",
  "requests",
  "tableprint",
  "wasabi",
  "yaml",
  "click",
]

[tool.black]
line-length = 88
target-version = ["py37"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist

)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
